Wquantiles.jl Documentation

Overview

This module contains functions to compute weighted quantiles of data series. One version of this, Wquantile takes advantage of parallelization. To use it make sure that you are running Julia with multiple threads.

There is an associated Jupyter notebook at /src/WquantileTest.ipynb.

Functions

Wquantiles.wquantile — Method

wquantile(x, w, q[; chk=true, norm_wgt=true, sort_q=true])

Finds the q weighted quantile values from the vector x.

Type Constraints

S <: Real

V <: Real

Arguments

x ::AbstractVector{T}: Vector(n) of values from which to find quantiles.

w ::AbstractVector{S}: Vector(n) of weights to use.

q ::AbstractVector{V}: Vector(l) of quantile values.

Keyword Arguments

chk=true :: Bool: If true, check the input contract described below.

norm_wgt=true:: Bool: If true, normalize the weights.

NOTE: If norm_wgt is false, it is ASSUMED that w is already normalized.

sort_q=true ::Bool: If true, sort the quantile vector, q.

NOTE: If sort_q is false, it ASSUMED that q is already sorted.

Input Contract

The type of x implements sortable.

|x| == |w| – Length of x matches length of weights.

∀i, w[i] >= 0 – The weights are non-negative.

Σ w[i] > 0 – The sum of the weights is positive.

∀i, q[i] <= 1 – The quantile values are in $[0,1]$.

∀i, q[i] >= 0

Return

The vector(l) of weighted quantile values from x. Letting qs be the sorted quantiles of q. The entry i is the $i^{\rm th}$ quantile (in qs) of x.

Wquantiles.Wquantile — Method

Wquantile(X, w, q; chk=true, norm_wgt=true, sort_q=true)

Finds the q weighted quantile values from the columns of the matrix X.

Type Constraints

S <: Real

V <: Real

Arguments

X ::AbstractMatrix{T}: Matrix(n,m) of values from which to find quantiles.

w ::AbstractVector{S}: Vector(n) of weights to use.

q ::AbstractVector{V}: Vector(l) of quantile values.

Keyword Arguments

chk=true ::Bool: If true, check the input contract described below.

norm_wgt=true::Bool: If true, normalize the weights.

NOTE: If norm_wgt is false, it is ASSUMED that w is already normalized.

sort_q=true ::Bool: If true, sort the quantile vector, q.

NOTE: If sort_q is false, it is ASSUMED that q is already sorted.

Input Contract

The type of X implements sortable.

∀i, |X[:, i]| == |w| – Length of each column of X matches length of weights.

∀i, w[i] >= 0 – Weights are non-negative.

Σ w[i] > 0 – The sum of the weights is positive.

∀i, q[i] <= 1 – The quantiles values are in $[0,1]$.

∀i, q[i] >= 0

Return

The (l,m) matrix of weighted quantile values from X. Letting qs be the sorted quantiles of q. The entry (i,j) is the $i^{\rm th}$ quantile (in qs) from the $j^{\rm th}$ column of X.

Wquantiles.wquantile — Method

wquantile(X, W, q; chk=true)

Finds the q weighted quantile values from the columns of the matrix X.

Type Constraints

S <: Real

V <: Real

Arguments

X ::AbstractMatrix{T}: Matrix(n,m) of values from which to find quantiles.

W ::AbstractMatrix{S}: Matrix(n,m) of weights to use.

q ::AbstractVector{V}: Vector(l) of quantile values.

Keyword Arguments

chk=true::Bool : If true, check the input contract described below.

Input Contract

The type of X is sortable.

size(X) == size(W)

∀i∀j, W[i, j] >= 0 – Weights are non-negative in each column.

∀i,Σ_j w[i,j] > 0 – The column sum of the weights is positive.

∀i, 0 <= q[i] <= 1 – The quantiles values are in $[0,1]$.

Return

The (l,m) matrix of weighted quantile values from X. Letting qs be the sorted quantiles of q. The entry (i,j) is the $i^{\rm th}$ quantile (in qs) from the $j^{\rm th}$ column of X.

Index

Wquantiles.Wquantile

Wquantiles.wquantile

Wquantiles.wquantile